app/vitals-processor/requirements.txt
functions-framework==3.*
google-auth>=2.0.0
requests>=2.28.0
